[Task-43] Updated Readme, Main, Hello, and Handler Functions
