+++ handler/handler.csv
@@ -4,3 +4,4 @@ ID,Name,Occupation
 3,Haris,Front End Developer
 4,Shahid,Back End Developer
 5,Shahzad,Full Stack Developer
+6,Ehtisham, Associate Researcher
