+++ handler/hello.go
@@ -7,12 +7,12 @@ import (
 	"time"
 )
 
-type Request struct {
+type HelloRequest struct {
 	Name string `json:"name"`
 	Path string
 }
 
-type Response struct {
+type HelloResponse struct {
 	Code      int    `json:"code"`
 	Message   string `json:"message"`
 	Timestamp string `json:"timestamp"`
@@ -21,14 +21,14 @@ type Response struct {
 func HelloHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	var req Request
+	var req HelloRequest
 	err := json.NewDecoder(r.Body).Decode(&req)
 	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		return 
 	}
 
-	response := &Response{
+	response := &HelloResponse{
 		Code:      http.StatusOK,
 		Message:   fmt.Sprintf("Welcome %s!", req.Name),
 		Timestamp: time.Now().UTC().Format(time.RFC3339),
